Consolidate feature package manifests
