--- Features/Settings/Package.resolved
@@ -1,15 +1,0 @@
-{
-  "originHash" : "0000000000000000000000000000000000000000000000000000000000000000",
-  "pins" : [
-    {
-      "identity" : "swiftsoup",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/scinfu/SwiftSoup.git",
-      "state" : {
-        "revision" : "3a439f9eccc391b264d54516ce640251552eb0c4",
-        "version" : "2.10.3"
-      }
-    }
-  ],
-  "version" : 3
-}
